+++ framework/validators/TrimValidator.php
@@ -7,6 +7,8 @@
 
 namespace yii\validators;
 
+use yii\helpers\Json;
+
 /**
  * This class converts the attribute value(s) to string(s) and strip characters.
  *
@@ -65,7 +67,7 @@ class TrimValidator extends Validator
         ValidationAsset::register($view);
         $options = $this->getClientOptions($model, $attribute);
 
-        return 'value = yii.validation.trim($form, attribute, ' . json_encode($options) . ', value);';
+        return 'value = yii.validation.trim($form, attribute, ' . Json::htmlEncode($options) . ', value);';
     }
 
     /**